CLI-842 Move the command API surfaces onto their own clients and retire SonarQubeClient
