+++ src/commands/analyze/sqaa-display-json.ts
@@ -20,12 +20,11 @@
 
 // JSON report builders for SQAA results.
 
-import type { SqaaAnalysisDepth, SqaaIssue } from '@/core/server/client.ts';
-
 import type { FileFailure, FileSuccess, RunTally } from './sqaa-analysis.ts';
 import { toRelativePosixPath } from './sqaa-api.ts';
 import type { IgnoredFile } from './sqaa-changeset.ts';
 import { type GlobalSqaaErrorKind, globalSqaaErrorKind } from './sqaa-errors.ts';
+import type { SqaaAnalysisDepth, SqaaIssue } from './sqaa-wire-types.ts';
 
 export interface SqaaJsonReport {
   files: Array<{
