+++ src/commands/analyze/sqaa-wire-types.ts
@@ -1,0 +1,74 @@
+/*
+ * SonarQube CLI
+ * Copyright (C) SonarSource Sàrl
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+// Request and response shapes of the Vortex analysis endpoint, exactly as they go over the
+// wire. A leaf module on purpose: it imports nothing, so every SQAA module can depend on it.
+
+export type SqaaAnalysisDepth = 'STANDARD' | 'DEEP';
+
+export type SqaaFileScope = 'MAIN' | 'TEST';
+
+export interface SqaaAnalysisFile {
+  path: string;
+  content: string;
+  scope?: SqaaFileScope;
+}
+
+export interface SqaaAnalysisRequest {
+  /** Cloud-only: the Server hub forces the request onto the instance's default organization. */
+  organizationKey?: string;
+  projectKey: string;
+  branchName?: string;
+  files: SqaaAnalysisFile[];
+  analysisDepth?: SqaaAnalysisDepth;
+}
+
+export interface SqaaAnalysisResponse {
+  id: string;
+  issues: SqaaIssue[];
+  patchResult?: {
+    newIssues: SqaaIssue[];
+    matchedIssues: SqaaIssue[];
+    closedIssues: string[];
+  } | null;
+  errors?: Array<{ code: string; message: string }> | null;
+}
+
+export interface SqaaIssue {
+  id: string;
+  filePath?: string | null;
+  message: string;
+  rule: string;
+  textRange?: {
+    startLine: number;
+    endLine: number;
+    startOffset: number;
+    endOffset: number;
+  } | null;
+  flows?: Array<{
+    type: string;
+    description?: string | null;
+    locations: Array<{
+      textRange?: { startLine: number; endLine: number } | null;
+      message?: string | null;
+      file?: string | null;
+    }>;
+  }> | null;
+}
