+++ tests/unit/commands/analyze/sqaa-display.test.ts
@@ -20,8 +20,8 @@
 
 import { afterEach, describe, expect, it } from 'bun:test';
 
+import type { SqaaIssue } from '@/commands/analyze/sqaa-wire-types.ts';
 import { CommandFailedError } from '@/core/command-error.ts';
-import type { SqaaIssue } from '@/core/server/client.ts';
 import { clearMockUiCalls, getMockUiCalls, setMockUi } from '@/core/ui';
 
 import type { FileResult, RunTally } from '../../../../src/commands/analyze/sqaa-analysis.ts';
