+++ tests/unit/commands/hook/agent-post-tool-use.test.ts
@@ -23,6 +23,7 @@ import { join } from 'node:path';
 
 import { afterEach, beforeEach, describe, expect, it, spyOn } from 'bun:test';
 
+import * as clientModule from '@/commands/analyze/sqaa-analysis-client.ts';
 import * as sqaaTelemetry from '@/commands/analyze/sqaa-analysis-telemetry.ts';
 import {
   SQAA_CLAUDE_POST_TOOL_USE_CALLER_COMMAND,
@@ -32,7 +33,6 @@ import * as authResolver from '@/core/auth/auth-resolver.ts';
 import { CommandInvocationContext } from '@/core/commands/invocation-context.ts';
 import * as processLib from '@/core/process/process.ts';
 import * as projectInfo from '@/core/project-info.ts';
-import * as clientModule from '@/core/server/client.ts';
 
 import { agentPostToolUse } from '../../../../src/commands/hook/agent-post-tool-use.ts';
 import { contextAugmentationPostToolUseSubscriber } from '../../../../src/commands/hook/context-augmentation-hook-subscriber.ts';
@@ -91,7 +91,7 @@ describe('agentPostToolUse', () => {
       configSources: [],
     });
     createAnalysisSpy = spyOn(
-      clientModule.SonarQubeClient.prototype,
+      clientModule.SqaaAnalysisClient.prototype,
       'createAnalysis',
     ).mockResolvedValue({ id: 'analysis-id', issues: [], errors: null });
     emitSqaaAnalysisTelemetrySpy = spyOn(
